refactored to use yarn and create-react-app bootstrap, upgraded packages to latest
