+++ .gitignore
@@ -13,7 +13,11 @@
 
 # misc
 .DS_Store
+.vscode
+
+.env
 .env.local
+.env.development
 .env.development.local
 .env.test.local
 .env.production.local
