--- index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="stylesheet" href="/style/style.css" />
-    <link
-      rel="stylesheet"
-      href="https://cdn.rawgit.com/twbs/bootstrap/48938155eb24b4ccdde09426066869504c6dab3c/dist/css/bootstrap.min.css"
-    />
-  </head>
-  <body class="root">
-    <div class="container"></div>
-  </body>
-  <script src="/bundle.js"></script>
-</html>
