+++ src/components/app.js
@@ -1,0 +1,52 @@
+import _ from 'lodash'
+
+import React, { Component } from 'react'
+import YTSearch from 'youtube-api-search'
+
+import SearchBar from './search_bar'
+import VideoList from './video_list'
+import VideoDetail from './video_detail'
+
+const API_KEY = process.env.REACT_APP_apiAccessKey;
+
+// Create a new component. This component will produce some HTML
+class App extends Component {
+    constructor(props) {
+        super(props)
+
+        this.state = {
+            videos: [],
+            selectedVideo: null
+        }
+
+        this.videoSearch('sea otter')
+    }
+
+    videoSearch(term) {
+        YTSearch({ key: API_KEY, term: term }, (videos) => {
+            this.setState({
+                videos: videos,
+                selectedVideo: videos[0]
+            })
+        })
+    }
+
+    render() {
+        const videoSearch = _.debounce((term) => {
+            this.videoSearch(term)
+        }, 300)
+
+        return (
+            <div>
+                <SearchBar onSearchTermChange={videoSearch} />
+                <VideoDetail video={this.state.selectedVideo} />
+                <VideoList
+                    onVideoSelect={(selectedVideo) => this.setState({ selectedVideo })}
+                    videos={this.state.videos}
+                />
+            </div>
+        )
+    }
+}
+
+export default App;
